desktop/core/ext-py/django-oidc-auth-0.0.11/requirements.txt
# Production requirements
Django>1.5
pyjwkest==0.6.2
requests==2.2.1
South==1.0.2

# Test requirements
django-nose==1.4
mock==1.0.1
tox
